rename flux central control to flux/main.js
